Wrap in AMD module and add Backbone.SimperiumModel.
Wrap the model in a define() statement to work well with require.js. Defaults to
using the global namespace if define() is not present. Note that if used as a
module it does require the Simperium JS code to be loaded first. That hasn't
been added as a dependency because the Simperium library is not yet AMD-enabled.

Also adds Backbone.SimperiumModel (props @fredrocious) to allow syncing a single
Model as well as a Collection.